#28: In-game quest requirements display. (not finished yet but network messages are being sent, and the empty "Quest Requirements" window is shown in-game.)
git-svn-id: svn://www.knightsgame.org.uk/trunk@81 00979ad9-a3a6-4f7a-b49c-25ba5319b333

+++ src/server/impl/server_mini_map.cpp
@@ -113,7 +113,7 @@ void ServerMiniMap::appendMiniMapCmds(std::vector<ubyte> &vec) const
             buf.writeUbyte(it->start_x);
             buf.writeUbyte(it->y);
             buf.writeUbyte(it->cols.size());
-            for (int i = 0; i < it->cols.size(); ++i) {
+            for (size_t i = 0; i < it->cols.size(); ++i) {
                 // TODO: could compress even more (each square would fit
                 // into 2 bits) but we'll just use one byte per square...
                 buf.writeUbyte(it->cols[i]);
